climbing: refactor - move onMove+onClick to RoutesLayer
